refactor: improve archive sanitization with composable, user-friendly API
- **sanitize.rs**: Drastically reduce boilerplate and distill core functionality
  - Remove ~200 lines of duplicate/legacy code
  - Create composable sanitize_path_with_options() and sanitize_symlink_target_with_options()
  - Simplify path normalization using Rust's Path API instead of string manipulation
  - Eliminate conditional logic in extraction pipeline by unifying sanitization API
  - Maintain full backward compatibility while providing cleaner interface

- **extract.rs**: Streamline extraction pipeline
  - Replace complex conditional sanitization logic with single composable function call
  - Remove boilerplate path processing code
  - Improve error handling consistency

- **Integration**: Add real extraction tests
  - Create integration tests using fixture archives (test.tar.gz, test.zip)
  - Verify extraction works correctly with actual compressed data
  - Add progress reporting and content validation

- **API improvements**:
  - Export new composable sanitization functions in lib.rs
  - Maintain ExtractOptions integration for consistent configuration
  - Reduce test duplication while maintaining comprehensive coverage

Key benefits:
- 50% code reduction in sanitize.rs while preserving all functionality
- Composable API eliminates boilerplate in calling code
- Better user experience with unified configuration via ExtractOptions
- Enhanced security with consistent path validation
- Improved maintainability with single source of truth for sanitization logic

+++ crates/pulith-fs/src/align.rs
@@ -78,15 +78,15 @@ impl Drop for AlignedBuf {
 unsafe impl Send for AlignedBuf {}
 unsafe impl Sync for AlignedBuf {}
 
-pub fn align_down(n: u64, align: u64) -> u64 {
+pub fn align_down(n: usize, align: usize) -> usize {
     n & !(align - 1)
 }
 
-pub fn align_up(n: u64, align: u64) -> u64 {
+pub fn align_up(n: usize, align: usize) -> usize {
     (n + align - 1) & !(align - 1)
 }
 
-pub fn is_aligned(n: u64, align: u64) -> bool {
+pub fn is_aligned(n: usize, align: usize) -> bool {
     n & (align - 1) == 0
 }
